Modificaciones en controlador Archivo
- Se crea metodo para modificar el archivo si este ya existe, actualizando su tamaño en la bd
- se añade el tamaño del archivo a la bd, asi como manejo de excepciones para posibles errores.
- se añade metodo para calcular el espacio usado por el usuario

+++ CloudDrive/grails-app/domain/clouddrive/Archivo.groovy
@@ -5,12 +5,13 @@ class Archivo {
     int id
     Usuario uid_usr
     String nombre
-    int size
+    Long size
     String file_path
     
 
     static constraints = {
         id(unique:true)
+        
     }
 
     static mapping = {
